Moved 'rectoverlap' operator to Ric.Operators. Added `rectoverlap` operator and related documentation; refactored operator registration
