+++ DocsExamples/FabInspector-Operators.md
@@ -18,7 +18,6 @@ FabInspector operators extend the [JSON Logic](https://json-everything.net/json-
 ## Contents
 
 - [REST API Operators](#rest-api-operators): `apiget`, `dfsget`, `daxquery`, `sqlquery`, `scannerapi`
-- [Layout & Geometry](#layout--geometry): `rectoverlap`
 
 ---
 
@@ -210,43 +209,6 @@ Calls the [Power BI Admin Workspace Info API](https://learn.microsoft.com/en-us/
 ```
 
 See also: [Example-scannerapi-rules.json](Example-scannerapi-rules.json)
-
----
-
-## Layout & Geometry
-
-### `rectoverlap`
-
-Detects overlapping rectangles in a list of named rectangular regions. Optionally expands each rectangle by a pixel margin before checking for overlaps. Returns the names of any rectangles that overlap with at least one other.
-
-| Parameter | Type | Description |
-|---|---|---|
-| input | array | Array of rectangle objects, each with integer properties `name`, `x`, `y`, `width`, `height` |
-| margin | number | Pixel amount to expand each rectangle on all sides before the overlap check (optional, default `0`) |
-
-**Returns:** Array of `name` values for rectangles that overlap with at least one other rectangle.
-
-```json
-{
-  "rectoverlap": [
-    {
-      "map": [
-        { "part": "Visuals" },
-        {
-          "torecord": [
-            "name",   { "var": "name" },
-            "x",      { "var": "visual.position.x" },
-            "y",      { "var": "visual.position.y" },
-            "width",  { "var": "visual.position.width" },
-            "height", { "var": "visual.position.height" }
-          ]
-        }
-      ]
-    },
-    5
-  ]
-}
-```
 
 ---
