+++ FabInspector.CLI/Program.cs
@@ -102,13 +102,13 @@ internal partial class Program
                 new SliceOperator(),
                 new NowOperator(),
                 new DateDiffOperator(),
-                new LetOperator()
+                new LetOperator(),
+                new RectangleOverlapOperator()
         }));
 
         registries.Add(new JsonLogicOperatorRegistry(
         new FabInspectorSerializerContext(),
         new IJsonLogicOperator[] {
-                new RectangleOverlapOperator(),
                 new DaxQueryOperator(),
             new SqlQueryOperator(),
                 new ApiGetOperator(),
