+++ FabInspector.Operators/FabInspectorSerializerContext.cs
@@ -3,7 +3,6 @@
 namespace FabInspector.Operators;
 
 [JsonSerializable(typeof(Json.Logic.Rule))]
-[JsonSerializable(typeof(RectOverlapRule))]
 [JsonSerializable(typeof(DaxQueryRule))]
 [JsonSerializable(typeof(SqlQueryRule))]
 [JsonSerializable(typeof(ApiGetRule))]
